+++ dev-packages/cli/templates/fc/admin-app/src/browser/module.ts
@@ -1,3 +1,4 @@
 import './user.view';
+import './application-lifecycle';
 import { autoBind } from '@malagu/core';
 export default autoBind();