+++ dev-packages/cli/templates/fc/admin-app/src/browser/user.view.tsx
@@ -1,16 +1,38 @@
 import * as React from 'react';
 import { View } from '@malagu/react';
-import { LoginUserManager } from '@malagu/shell/lib/browser';
-import { ContainerUtil } from '@malagu/core';
+import { RpcUtil } from '@malagu/rpc';
+import { DataTable } from 'grommet';
+import { UserService, User } from '../common';
 
 function Users() {
-    const loginUserManager = ContainerUtil.get<LoginUserManager>(LoginUserManager);
-    loginUserManager.userSubject.next({
-        id: 'kevin',
-        name: 'Kevin',
-        avatar: '//s.gravatar.com/avatar/b7fb138d53ba0f573212ccce38a7c43b?s=80'
-    });
-    return (<div></div>);
+
+    const [ data, setData ] = React.useState<User[]>([]);
+
+    React.useEffect(() => {
+        const userService = RpcUtil.get<UserService>(UserService);
+        userService.list().then(users => setData(users))
+
+    }, [])
+    
+    return (
+        <DataTable
+            margin={{ vertical: 'medium' }}
+            columns={[
+                {
+                    property: 'name',
+                    header: 'Name',
+                    primary: true
+                },
+                {
+                    property: 'age',
+                    header: 'Age'
+                }
+            ]}
+            data={data}
+        >
+
+        </DataTable>
+    );
 }
 
 @View({ component: Users})