+++ dev-packages/cli/templates/fc/admin-app/src/node/entity/user.ts
@@ -1,0 +1,15 @@
+import { Entity, PrimaryGeneratedColumn, Column } from 'typeorm';
+import { User as IUser } from '../../common';
+
+@Entity()
+export class User implements IUser {
+    
+    @PrimaryGeneratedColumn()
+    id: number;
+
+    @Column()
+    name: string;
+
+    @Column()
+    age: number;
+}