+++ dev-packages/cli/templates/fc/admin-app/src/node/module.ts
@@ -1,3 +1,8 @@
-import './welcome-server';
+import './user-service';
+import * as entities from './entity';
 import { autoBind } from '@malagu/core';
+import { autoBindEntities } from '@malagu/typeorm';
+
+autoBindEntities(entities)
+
 export default autoBind();